requirements.txt
Flask
Flask-SQLAlchemy
mysql-connector-python
MarkupSafe
nltk
transformers
eventlet
Flask-Cors
Flask-SocketIO
Flask-APScheduler
python-dotenv
numpy
